refactor(UBA): sideload HRE environment variable
This change essentially side-loads the HRE variables to prevent the need to directly interface with the Hardhat Runtime Environment.
