+++ deploy/003_deploy_optimism_spokepool.ts
@@ -1,4 +1,4 @@
-import { deployNewProxy } from "../utils";
+import { deployNewProxy } from "../utils/utils.hre";
 import { DeployFunction } from "hardhat-deploy/types";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
