+++ test/chain-adapters/Optimism_Adapter.ts
@@ -1,6 +1,15 @@
 import { sampleL2Gas, amountToLp, mockTreeRoot, refundProposalLiveness, bondAmount } from "./../constants";
-import { ethers, expect, Contract, FakeContract, SignerWithAddress, createFake, hre } from "../../utils/utils";
-import { getContractFactory, seedWallet, randomAddress } from "../../utils/utils";
+import {
+  ethers,
+  expect,
+  Contract,
+  FakeContract,
+  SignerWithAddress,
+  createFake,
+  getContractFactory,
+  seedWallet,
+  randomAddress,
+} from "../../utils/utils";
 import { hubPoolFixture, enableTokensForLP } from "../fixtures/HubPool.Fixture";
 import { constructSingleChainTree } from "../MerkleLib.utils";
