+++ test/fixtures/SpokePool.Fixture.ts
@@ -1,12 +1,6 @@
-import {
-  getContractFactory,
-  SignerWithAddress,
-  Contract,
-  hre,
-  ethers,
-  BigNumber,
-  defaultAbiCoder,
-} from "../../utils/utils";
+import { getContractFactory, SignerWithAddress, Contract, ethers, BigNumber, defaultAbiCoder } from "../../utils/utils";
+import { hre } from "../../utils/utils.hre";
+
 import * as consts from "../constants";
 
 export const spokePoolFixture = hre.deployments.createFixture(async ({ ethers }) => {
