+++ test/merkle-distributor/MerkleDistributor.ts
@@ -1,16 +1,6 @@
 /* eslint-disable no-unused-expressions */
 
-import {
-  ethers,
-  getContractFactory,
-  SignerWithAddress,
-  Contract,
-  toWei,
-  toBN,
-  expect,
-  keccak256,
-  defaultAbiCoder,
-} from "../../utils/utils";
+import { ethers, getContractFactory, SignerWithAddress, Contract, toWei, toBN, expect } from "../../utils/utils";
 import { deployErc20 } from "../gas-analytics/utils";
 import { MAX_UINT_VAL, MerkleTree } from "@uma/common";
